removing examples and jsdoc
Example has been removed for now, as it's technically in the app/
directory now. A new example should be added to demonstrate the use of
the app as a simple server that can be launched.

jsdoc has been removed in favor of using the Sphinx documentation,
although the comments are still in the code and _should_ be added
anyway.